% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{match_arg}
\alias{match_arg}
\title{match_arg is a better match.arg}
\usage{
match_arg(arg, choices, several.ok = FALSE)
}
\description{
A simple wrapper around match.arg that returns a slightly
better error message when used programmatically.
}

