% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoenc.R
\name{predict.autoenc}
\alias{predict.autoenc}
\title{Predict method for a fitted autoencoder}
\usage{
\method{predict}{autoenc}(object, newdata, ...)
}
\arguments{
\item{object}{autoenc. An object returned by \code{sparse_autoenc} function.}

\item{newdata}{dataframe or matrix. An object to predict upon}

\item{\dots}{further arguments passed to or from other methods}
}
\value{
Returns a numeric matrix, where each row of the input matrix corresponds to
  the value at the last hidden layer of the autoencoder.
}
\description{
Predict method for a fitted autoencoder
}

