% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/param_validation.R
\name{validate_autoenc_params}
\alias{validate_autoenc_params}
\title{Validate and preprocess inputs to autoenc call}
\usage{
validate_autoenc_params(...)
}
\description{
Converts the params from autoenc inputs into an
intermediate representation that will be used by modeling functions.
}

